+++ src/console/tabs/agents-tab.js
@@ -227,7 +227,7 @@ ${_tl('bmadDesc')}
 
   const box = blessed.box({
     parent: screen,
-    top: 4,
+    top: 5,
     left: 0,
     width: '100%',
     bottom: 2,