+++ src/console/tabs/help-tab.js
@@ -160,7 +160,7 @@ export function createHelpTab(screen, services) {
 
   const box = blessed.box({
     parent: screen,
-    top: 4,
+    top: 5,
     left: 0,
     width: '100%',
     bottom: 2,