+++ src/console/tabs/install-tab.js
@@ -166,7 +166,7 @@ export function createInstallTab(screen, services) {
 
   const box = blessed.box({
     parent: screen,
-    top: 4,
+    top: 5,
     left: 0,
     width: '100%',
     bottom: 2,